Junit new version

+++ com/elections/src/main/java/com/dan323/elections/simulations/VotingSimulation.java
@@ -5,5 +5,5 @@ import java.util.Map;
 @FunctionalInterface
 public interface VotingSimulation {
 
-    Map<String,Integer> getVotes();
+    Map<String,Long> getVotes();
 }
