+++ com/elections/src/main/java/com/dan323/elections/systems/quo/Quota.java
@@ -4,6 +4,6 @@ import java.util.Map;
 import java.util.function.BiFunction;
 
 @FunctionalInterface
-public interface Quota<T> extends BiFunction<Map<T,Integer>, Integer, Double> {
+public interface Quota<T> extends BiFunction<Map<T,Long>, Integer, Double> {
 
 }
